pyproject.toml
[tool.poetry]
name = "gd-task-2"
version = "0.1.0"
description = ""
authors = ["Alexey Yamushev <Privet.02@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.111.1"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
pyjwt = "^2.8.0"
bcrypt = "^4.1.3"
asyncpg = "^0.29.0"
slowapi = "^0.1.9"
passlib = "^1.7.4"
redis = "^5.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
